Add Gun-Vue project
I've configured the deploy to skynet script and will keep the Skynet version of https://gun-vue.js.org in sync. It's a collection of UI patterns to be implemented in p2p GunDB apps, that may build upon this lib. But it itself may become a useful app for many basic cases. Like encrypted messaging and structured data sharing.